'photo' module
